Use local-hostname for HELO command
In the SMTP-HANDSHAKE function, a EHLO command is sent to the server if ESMTP
features are required and HELO otherwise.

This function also takes a keyword argument, LOCAL-HOSTNAME, that should be
used as the argument to EHLO/HELO. However, this did not work collectly.

While EHLO commands were followed by the value of LOCAL-HOSTNAME, HELO
commands, on the other hand, completely disregarded LOCAL-HOSTNAME, instead
calling (USOCKET::GET-HOST-NAME) directly. In other words, overriding the
hostname sent in HELO commands did not work.

This commit ensures that the value of LOCAL-HOSTNAME is always used in HELO or
EHLO commands, thereby ensuring that the HELO/EHLO hostnames can always be
overridden by users.

Note: this function seems to assume that LOCAL-HOSTNAME is set to a non-NIL
value. This might not always be the case and there are ways to cause “HELO
NIL” or “EHLO NIL” to be sent to the SMTP server. This latent issue will be
addressed in a follow-up commit.