Make the X-Mailer header suppressable
In some scenarios, it may be desirable to suppress the X-Mailer header in
e-mails completely, because revealing the name and versions of the software used
to send e-mails may be a security concern.

This patch enables suppressing the X-Mailer header by binding
CL-SMTP::*X-MAILER* to NIL, and adds documentation to that effect.