Update CL-SMTP:SEND-EMAIL documentation in README
The documentation of CL-SMTP:SEND-EMAIL in the README file was out of sync
with the code and needed an update.

The lambda list of that function appearing in the README is updated to be
identical to that in the code.

The arguments’ descriptions are reordered to be in the same order as the
arguments in the function’s lambda list. A description for the LOCAL-HOSTNAME
keyword parameter was also missing.