Always provide a hostname to SMTP-HANDSHAKE
One code path in the SMTP-HANDSHAKE function could cause NIL to be sent as a
hostname in a HELO or EHLO command.

Consider the following code typed into a REPL:

  (cl-smtp:with-smtp-mail (s "my.mail.server.example" "me@example.com"
                             '("you@example.com"))
    (cl-smtp:write-rfc8822-message s "me@example.com" '("you@example.com")
                                   "Testing CL-SMTP" "Oh hi there"))

(The ESMTP case can be tested by adding :SSL :STARTTLS to the call to
WITH-SMTP-MAIL.)

Because no :LOCAL-HOSTNAME keyword argument was supplied to WITH-SMTP-MAIL,
eventually, SMTP-HANDSHAKE was called with NIL as a :LOCAL-HOSTNAME keyword
parameter. And because no checks were done in the case SMTP-HANDSHAKE needs to
negociate ESMTP features, the EHLO command was sent with NIL as hostname.

It is better to substitute defaults for values such as the local hostname in
functions that are exported to users: doing so lets dependent, internal
functions such as SMTP-HANDSHAKE stay simple. SEND-EMAIL already does this, so
the same simply needs to be done with WITH-SMTP-MAIL too.